Turn source image project into a combobox.
The previous select was too restrictive as it didn't allow to provide a value for this field other than the project where the VM is to be scheduled.

Source image could be located in a different project and shared to the project where the VM is scheduled.

Also removed usage of anonymous inner class when a simpler alternative is available.